+++ src/libs/conduit/conduit_data_array.cpp
@@ -320,6 +320,32 @@ DataArray<T>::diff_compatible(const DataArray<T> &array, Node &info, const float
 }
 
 //---------------------------------------------------------------------------//
+template <typename T>
+std::string
+DataArray<T>::to_string(const std::string &protocol) const
+{
+    if(protocol != "json")
+    {
+        // unsupported
+        CONDUIT_ERROR("Unknown DataArray::to_string protocol:" << protocol
+                     <<"\nSupported protocols:\n" 
+                     <<" json");
+    }
+
+    return to_json(); 
+}
+
+
+//---------------------------------------------------------------------------//
+template <typename T>
+std::string
+DataArray<T>::to_string_default() const
+{ 
+    return to_string();
+}
+
+
+//---------------------------------------------------------------------------//
 template <typename T> 
 std::string
 DataArray<T>::to_json() const