+++ src/libs/conduit/conduit_data_type.hpp
@@ -525,7 +525,14 @@ public:
 //-----------------------------------------------------------------------------
 // Transforms
 //-----------------------------------------------------------------------------
-    std::string         to_json() const;  
+    std::string         to_string(const std::string &protocol="json") const;
+
+    // NOTE(cyrush): The primary reason this function exists is to enable 
+    // easier compatibility with debugging tools (e.g. totalview, gdb) that
+    // have difficulty allocating default string parameters.
+    std::string         to_string_default() const;
+    
+    std::string         to_json() const;
     void                to_json_stream(std::ostream &os) const;
 
     void                compact_to(DataType &dtype) const;